Allow middleware config for access token POST call
Allow for a :middleware key in the profile configuration, where one can
override the clj-http middleware stack. This is for cases where the
targeted endpoint eg. sends back badly formatted cookies. In such a
case removing the wrap-cookies middleware (which is not needed anyway)
prevents the get-access-token post call from crashing.